Simplify the rmf_owl body frame
The body frame now coincides with the model frame. Moved the collision
box so that its center coincides with the center of the visualization
mesh, making it the same as the mesh AABB. Moved the rmf_owl starting
position 10 cm up so that it starts above ground. Moved the rotors so
that they match the moved mesh.