Takeoff to wait for OKVIS2 to be running
When running with the neural network, OKVIS2 takes longer to start
producing odometry estimates as the network has got to "warm up". I the
script has finished before OKVIS2 is running, the drone will not take
off. Made a change so the script waits for the first okvis osometry
message before arm and takeoff commands