[BUGFIX] removed blocky exposure artefacts
There is a bug in the current gz sim, where activating the <extrinsics> tag for a sensor, blocky exposure artefacts are introduced. This fix removes the tag and adjusts the okvis config to represent the perfect extrinsics resulting from having perfect center points and focal lengths.

Feel free to undo this commit once the bug in gz is fixed